+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Bugfix: `find_subtype` failed when reads from both HIV and HCV were present.
 - Bugfix: when no RAS mutation is found, `reportdrm` failed.
 - Bugfix: when the frequency of a mutation is 0.5/0.5, `cns_max_freq` was longer than the reference.
 - Problems with an assert when all reads have indels: removed the assert, `phase_mutations` has limitations.